luks: add initrdUnlock option to luks type
This option, when set explicitly to 'false' prevents disko
from modifying the initrd configuration, and stops the
system from trying to unlock the disk at early boot time.

This means that for non boot-critical drives, they can be
lazily unlocked with a crypttab or other means later in the
boot process.