disko cli: fix error about impurity with nix
The script is generated by calling `nix-build` on the `cli.nix` file.
This fails on systems, which have the nix option pure_eval set, as nix
refuses to evaluate an untracked file. This would normally be impure,
but in this case the file is supplied in the same nix store path, i.e.
it's not referencing foreign files. Thus adding the `--impure` argument
should leave the nix call in fact still pure.