types: add installMountpoint to filesystem
Add an option to specify a different mountpoint for a filesystem during
installation.

This can be useful when using the impermanence NixOS module with /nix
on the persistent filesystem. During installation, we want to mount the
filesystem to / and write to /nix under it, whereas in the NixOS
configuration we want to mount the filesystem to /persistent and create
a bind mount from /nix to /persistent/nix.