disko2: Move to error codes and dedicated error rendering
This splits the concern of how to display errors off from the business
logic. This should make it easier to test functionality and allow us
to concentrate on how to actually display and communicate errors to the
user. At some point this might also make i18n feasible, because you only
have to swap out a single file.

This hasn't quite fixed the debug/trace story yet. It should be possible
to simply insert a "debug" command into any pipeline and see the context
and state at that point.