disko2: Add code generator for disko config type
There are still two issues: The type of "topology" in zpool is not
created, and gpt_partitions_options_hybrid_options for some reason
contains a `_create: "_create"` entry. This is an issue with the nix
evaluation, though, not with the code generator.

I'm fixing these issues manually to have some state I can start working
from.